refactor: Cria pasta modules/shortUrls
